Parse example classes from String instead of Groovy file in acceptance test. Split CompileStatic test cases to a separated file.
Parsing groovy files in acceptance test is making debugging hard as the groovy files will get compiled by Gradle and fail before the test is running. It is also hard to see what the test cases are testing by just looking at it when the parsed classes are kept in other files.